Docker Setup: Make install fails on fresh system
Changes:
	1) Checking the /opt/intel/iei path exists before uninstalling IEI

+++ docker_setup/setup_iei.py
@@ -127,4 +127,5 @@ if __name__ == '__main__':
         print("***********Installation Finished***********")
 
     if args.uninstall_iei:
-        uninstall_iei()
+        if os.path.exists(INSTALL_PATH):
+            uninstall_iei()
